修复营业额显示不出，并且修改为营业额只查询已入账的订单记录金额

+++ src/main/java/com/htkapp/modules/merchant/pay/service/serviceImpl/OrderRecordServiceImpl.java
@@ -378,7 +378,9 @@ public class OrderRecordServiceImpl implements OrderRecordService {
             if (result == null) {
                 return 0;
             } else {
-                return result.shortValue();
+                //数值太小显示不出营业额的原因
+//                return result.shortValue();
+                return result.doubleValue();
             }
         } catch (Exception e) {
             throw new Exception(Globals.CALL_DATABASE_ERROR);
